Move CudaSubsetMapping
